Refactoring connection handling to its own class
Added missing shut_down behavior
